Create README.md
This is the code for the Calorie Calculator for Activity 8. The original issue I was having, was when I ran the program, the form was completely blank. No button, textboxes, or labels. Now that isn't happening, but what is happening, is that when I run the program, and enter in the numbers for Fat Grams and Carb Grams, nothing displays in the portion calorie portions.